+++ codedoor/static/css/base.css
@@ -1,3 +1,88 @@
-body {
-	background-color: lightblue;
+@font-face {
+  font-family: 'Roboto';
+  src: url('/static/fonts/Roboto-Regular.ttf');
 }
+@font-face {
+  font-family: 'Roboto-Light';
+  src: url('/static/fonts/Roboto-Light.ttf');
+}
+@font-face {
+  font-family: 'Roboto-Medium';
+  src: url('/static/fonts/Roboto-Medium.ttf');
+}
+@font-face {
+  font-family: 'Roboto-Bold';
+  src: url('/static/fonts/Roboto-Bold.ttf');
+}
+
+ul,
+li,
+a,
+button,
+input,
+h1,
+h2,
+h3,
+h4,
+h5,
+h6,
+p,
+img,
+image,
+svg,
+path,
+g,
+canvas {
+	box-sizing: border-box;
+	background-color: transparent;
+	border: none;
+	text-decoration: none;
+	user-select: none;
+	list-style: none;
+	font-family: 'Roboto';
+	color: #7F7F7F;
+}
+
+* {
+    margin: 0;
+    padding: 0;
+}
+
+.container {
+    width: 600px;
+    margin: 0 auto;
+}
+
+@media (min-width: 768px) {
+    .container {
+        width: 768px;
+    }
+}
+@media (min-width: 992px) {
+    .container {
+        width: 992px;
+    }
+}
+@media (min-width: 1200px) {
+    .container {
+        width: 1170px;
+    }
+}
+
+.content {
+    margin: 40px 0 80px;
+}
+
+h1, h2 {
+    font-family: 'Roboto-Bold';
+    color: #484848;
+}
+h3, h4 {
+    font-family: 'Roboto-Medium';
+    color: #484848;
+}
+
+input {
+    color: #484848;
+}
+
