--- codedoor/static/css/baselegit.css
@@ -1,46 +1,0 @@
-body {
-
-}
-
-
-h1{ margin: 10px;
-	font-family: 'Anton', sans-serif;
-	color: #273A40;
-	font-size: 40px;
-}
-
-.centered {
-position: absolute;
-width:20%;
-left:40%;
-border-radius: 5px;
-border-style:solid;
-border-width: 2px;
-text-align: center;
-padding: 2%;
-top: 10%;
-}
-
-h2{
-	font-family: 'Anton', sans-serif;
-	color: #273A40;
-
-
-
-
-
-}
-p{
-   font-family: 'Magra', sans-serif;
-
-}
-
-a{
-	color: #273A40;
-
-}
-
-#logo1{
-  background-color: lightgray;
-
-}
