+++ codedoor/templates/codedoor/base.html
@@ -4,18 +4,29 @@
 <html>
 <head>
 	<title>{% block titletext %} {% endblock %}</title>
-	<link rel="stylesheet" type="text/css" href="{% static 'css/baselegit.css' %}">
+	<link rel="stylesheet" type="text/css" href="{% static 'css/base.css' %}">
+	<link rel="stylesheet" type="text/css" href="{% static 'css/nav.css' %}">
 	<link href="https://fonts.googleapis.com/css?family=Anton|Bree+Serif|Julius+Sans+One|Magra" rel="stylesheet">
 	
 </head>
 <body>
-	<div class="navbar">
+	<div class="nav">
 		<div class="container">
+			<div class="nav-item">
+				<a href="{% url 'codedoor:home' %}">
+					<img src="{% static 'images/logo.svg' %}" width="41px" height="58px" />
+				</a>
+			</div>
+			<div class="nav-item">
+				<img src="{% static 'images/profpic_circle.png' %}" />
+			</div>
 		</div>
 	</div>
-	<div class="container">
-		{% block content %}
-		{% endblock content %}
+	<div class="content">
+		<div class="container">
+			{% block content %}
+			{% endblock content %}
+		</div>
 	</div>
 </body>
 </html>
