+++ codedoor/templates/codedoor/createprofile.html
@@ -1,7 +1,6 @@
 {% extends 'codedoor/base.html' %}
 {% load static %}
 {% block content %}
-<html>
 <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
 <link rel="stylesheet" type="text/css" href="{% static 'css/createprofile.css' %}">
 <h1>
@@ -47,5 +46,4 @@
         <input type="submit" value="Submit">
     </div>
 </form>
-</html>
 {% endblock %}
