+++ codedoor/urls/applications.py
@@ -1,0 +1,7 @@
+from django.contrib.auth import views as auth_views
+from django.conf.urls import url
+from django.urls import path
+
+import codedoor.views as views
+
+urlpatterns = []
