+++ codedoor/views/__init__.py
@@ -1,0 +1,5 @@
+from .applications import *
+from .companies import *
+from .profiles import *
+from .questions import *
+from .root import *
