+++ codedoor/views/applications.py
@@ -1,0 +1,3 @@
+from django.shortcuts import render, redirect, get_object_or_404
+from django.http import HttpResponse
+from codedoor.models import Profile, Company, Question, Application
