+++ codedoor/views/root.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render, redirect, get_object_or_404
+from django.http import HttpResponse
+
+def home(request):
+    return HttpResponse("Hello!")
