+++ mysite/urls.py
@@ -15,9 +15,11 @@ Including another URLconf
 """
 from django.conf.urls import url, include
 from django.contrib import admin
+from django.shortcuts import redirect
 
 
 urlpatterns = [
+    url(r'^$', lambda r: redirect('codedoor:home')),
     url(r'^admin/', admin.site.urls),
     url(r'^codedoor/', include('codedoor.urls', namespace='codedoor'))
 ]
